+++ src/models/model.c
@@ -777,117 +777,129 @@ void model_init()
                 device_add(models[model]->device);
 }
 
-MODEL m_1 = {"[8088] AMI XT clone", ROM_AMIXT, "amixt", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 64, 640, 64, xt_init, NULL};
+/* 8088 PC's */
+MODEL m_amixt = {"[8088] AMI XT clone", ROM_AMIXT, "amixt", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 64, 640, 64, xt_init, NULL};
+MODEL m_ataripc3 = {"[8088] Atari PC3", ROM_ATARIPC3, "ataripc3", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 64, 640, 64, xt_init, NULL};
+MODEL m_cbm_pc10 = {"[8088] Commodore PC-10", ROM_CBM_PC10, "cbm_pc10", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 640, 640, 64, xt_init, NULL};
+MODEL m_compaq_pip = {"[8088] Compaq Portable Plus", ROM_COMPAQ_PIP, "compaq_pip", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 128, 640, 64, compaq_xt_init, NULL};
+MODEL m_dtk = {"[8088] DTK XT clone", ROM_DTKXT, "dtk", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 64, 640, 64, xt_init, NULL};
+MODEL m_genxt = {"[8088] Generic XT clone", ROM_GENXT, "genxt", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 32, 704, 16, xt_init, NULL};
+MODEL m_hyundaixt = {"[8088] Hyundai Super16T", ROM_SUPER16T, "hyundaixt", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 640, 640, 64, xt_init, NULL};
+MODEL m_hyundaixte = {"[8088] Hyundai Super16TE", ROM_SUPER16TE, "hyundaixte", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 640, 640, 64, xt_init, NULL};
+MODEL m_ibmpc = {"[8088] IBM PC", ROM_IBMPC, "ibmpc", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 64, 640, 32, xt_init, NULL};
+MODEL m_ibmpcjr = {"[8088] IBM PCjr", ROM_IBMPCJR, "ibmpcjr", {{"", cpus_pcjr}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED, 128, 640, 64, pcjr_init, &pcjr_device};
+MODEL m_ibmxt = {"[8088] IBM XT", ROM_IBMXT, "ibmxt", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 64, 640, 64, xt_init, NULL};
+MODEL m_jukopc = {"[8088] Juko XT clone", ROM_JUKOPC, "jukopc", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 64, 640, 64, xt_init, NULL};
+MODEL m_ledge_modelm = {"[8088] Leading Edge Model M", ROM_LEDGE_MODELM, "ledge_modelm", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 128, 704, 64, xt_init, NULL};
+MODEL m_ncr_pc4i = {"[8088] NCR PC4i", ROM_NCR_PC4I, "ncr_pc4i", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 256, 640, 64, xt_init, NULL};
+MODEL m_pxxt = {"[8088] Phoenix XT clone", ROM_PXXT, "pxxt", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 64, 640, 64, xt_init, NULL};
+MODEL m_europc = {"[8088] Schneider EuroPC", ROM_EUROPC, "europc", {{"", cpus_europc}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 512, 640, 128, europc_init, NULL};
+MODEL m_tandy = {"[8088] Tandy 1000", ROM_TANDY, "tandy", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED, 128, 640, 128, tandy1k_init, &tandy1000_device};
+MODEL m_tandy1000hx = {"[8088] Tandy 1000 HX", ROM_TANDY1000HX, "tandy1000hx", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED, 256, 640, 128, tandy1k_init, &tandy1000hx_device};
+MODEL m_to16_pc = {"[8088] Thomson TO16 PC", ROM_TO16_PC, "to16_pc", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 512, 640, 128, xt_init, NULL};
+MODEL m_t1000 = {"[8088] Toshiba T1000", ROM_T1000, "t1000", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED, 512, 1280, 768, xt_t1000_init, NULL};
+MODEL m_ltxt = {"[8088] VTech Laser Turbo XT", ROM_LTXT, "ltxt", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 64, 1152, 64, xt_laserxt_init, NULL};
+MODEL m_xi8088 = {"[8088] Xi8088", ROM_XI8088, "xi8088", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PS2, 64, 1024, 128, xt_xi8088_init, &xi8088_device};
+MODEL m_zdsupers = {"[8088] Zenith Data SupersPort", ROM_ZD_SUPERS, "zdsupers", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 128, 640, 128, xt_zenith_init, NULL};
 
-MODEL m_2 = {"[8088] Atari PC3", ROM_ATARIPC3, "ataripc3", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 64, 640, 64, xt_init, NULL};
-MODEL m_cbmPC10 = {"[8088] Commodore PC-10", ROM_CBM_PC10, "cbm_pc10", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 640, 640, 64, xt_init, NULL};
-MODEL m_3 = {"[8088] Compaq Portable Plus", ROM_COMPAQ_PIP, "compaq_pip", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 128, 640, 64, compaq_xt_init, NULL};
-MODEL m_4 = {"[8088] DTK XT clone", ROM_DTKXT, "dtk", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 64, 640, 64, xt_init, NULL};
-MODEL m_5 = {"[8088] Generic XT clone", ROM_GENXT, "genxt", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 32, 704, 16, xt_init, NULL};
-MODEL m_huyndai16t = {"[8088] Huyndai Super16T", ROM_SUPER16T, "hyundaixt", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 640, 640, 64, xt_init, NULL};
-MODEL m_huyndai16te = {"[8088] Huyndai Super16TE", ROM_SUPER16TE, "huyndaixte", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 640, 640, 64, xt_init, NULL};
-MODEL m_6 = {"[8088] IBM PC", ROM_IBMPC, "ibmpc", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 64, 640, 32, xt_init, NULL};
-MODEL m_7 = {"[8088] IBM PCjr", ROM_IBMPCJR, "ibmpcjr", {{"", cpus_pcjr}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED, 128, 640, 64, pcjr_init, &pcjr_device};
-MODEL m_8 = {"[8088] IBM XT", ROM_IBMXT, "ibmxt", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 64, 640, 64, xt_init, NULL};
-MODEL m_9 = {"[8088] Juko XT clone", ROM_JUKOPC, "jukopc", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 64, 640, 64, xt_init, NULL};
-MODEL m_10 = {"[8088] Leading Edge Model M", ROM_LEDGE_MODELM, "ledge_modelm", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 128, 704, 64, xt_init, NULL};
-MODEL m_11 = {"[8088] NCR PC4i", ROM_NCR_PC4I, "ncr_pc4i", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 256, 640, 64, xt_init, NULL};
-MODEL m_12 = {"[8088] Phoenix XT clone", ROM_PXXT, "pxxt", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 64, 640, 64, xt_init, NULL};
-MODEL m_13 = {"[8088] Schneider EuroPC", ROM_EUROPC, "europc", {{"", cpus_europc}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 512, 640, 128, europc_init, NULL};
-MODEL m_14 = {"[8088] Tandy 1000", ROM_TANDY, "tandy", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED, 128, 640, 128, tandy1k_init, &tandy1000_device};
-MODEL m_15 = {"[8088] Tandy 1000 HX", ROM_TANDY1000HX, "tandy1000hx", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED, 256, 640, 128, tandy1k_init, &tandy1000hx_device};
-MODEL m_16 = {"[8088] Thomson TO16 PC", ROM_TO16_PC, "to16_pc", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 512, 640, 128, xt_init, NULL};
-MODEL m_17 = {"[8088] Toshiba T1000", ROM_T1000, "t1000", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED, 512, 1280, 768, xt_t1000_init, NULL};
-MODEL m_18 = {"[8088] VTech Laser Turbo XT", ROM_LTXT, "ltxt", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 64, 1152, 64, xt_laserxt_init, NULL};
-MODEL m_19 = {"[8088] Xi8088", ROM_XI8088, "xi8088", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PS2, 64, 1024, 128, xt_xi8088_init, &xi8088_device};
-MODEL m_20 = {"[8088] Zenith Data SupersPort", ROM_ZD_SUPERS, "zdsupers", {{"", cpus_8088}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 128, 640, 128, xt_zenith_init, NULL};
+/* 8086 PC's */
+MODEL m_pc1512 = {"[8086] Amstrad PC1512", ROM_PC1512, "pc1512", {{"", cpus_pc1512}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED | MODEL_AMSTRAD, 512, 640, 128, ams_init, &ams1512_device};
+MODEL m_pc1640 = {"[8086] Amstrad PC1640", ROM_PC1640, "pc1640", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_HW | MODEL_AMSTRAD, 640, 640, 0, ams_init, &ams1512_device};
+MODEL m_pc2086 = {"[8086] Amstrad PC2086", ROM_PC2086, "pc2086", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_HW | MODEL_AMSTRAD, 640, 640, 0, ams_init, &ams2086_device};
+MODEL m_pc3086 = {"[8086] Amstrad PC3086", ROM_PC3086, "pc3086", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_HW | MODEL_AMSTRAD, 640, 640, 0, ams_init, &ams3086_device};
+MODEL m_pc5086 = {"[8086] Amstrad PC5086", ROM_PC5086, "pc5086", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_PS2, 640, 640, 0, pc5086_init, &f82c710_upc_device};
+MODEL m_ppc512 = {"[8086] Amstrad PPC512/640", ROM_PPC512, "ppc512", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_HW | MODEL_AMSTRAD, 512, 640, 128, ams_init, &ams1512_device};
+MODEL m_deskpro = {"[8086] Compaq Deskpro", ROM_DESKPRO, "deskpro", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 128, 640, 128, compaq_xt_init, NULL};
+MODEL m_olivetti_m24 = {"[8086] Olivetti M24", ROM_OLIM24, "olivetti_m24", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED | MODEL_OLIM24, 128, 640, 128, olim24_init, NULL};
+MODEL m_pc200 = {"[8086] Sinclair PC200", ROM_PC200, "pc200", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_HW | MODEL_AMSTRAD, 512, 640, 128, ams_init, &ams1512_device};
+MODEL m_tandy1000sl2 = {"[8086] Tandy 1000 SL/2", ROM_TANDY1000SL2, "tandy1000sl2", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED, 512, 768, 128, tandy1ksl2_init, NULL};
+MODEL m_t1200 = {"[8088] Toshiba T1200", ROM_T1200, "t1200", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED, 1024, 2048, 1024, xt_t1200_init, NULL};
+MODEL m_lxt3 = {"[8086] VTech Laser XT3", ROM_LXT3, "lxt3", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 512, 1152, 128, xt_laserxt_init, NULL};
 
-MODEL m_21 = {"[8086] Amstrad PC1512", ROM_PC1512, "pc1512", {{"", cpus_pc1512}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED | MODEL_AMSTRAD, 512, 640, 128, ams_init, &ams1512_device};
-MODEL m_22 = {"[8086] Amstrad PC1640", ROM_PC1640, "pc1640", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_HW | MODEL_AMSTRAD, 640, 640, 0, ams_init, &ams1512_device};
-MODEL m_23 = {"[8086] Amstrad PC2086", ROM_PC2086, "pc2086", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_HW | MODEL_AMSTRAD, 640, 640, 0, ams_init, &ams2086_device};
-MODEL m_24 = {"[8086] Amstrad PC3086", ROM_PC3086, "pc3086", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_HW | MODEL_AMSTRAD, 640, 640, 0, ams_init, &ams3086_device};
-MODEL m_25 = {"[8086] Amstrad PC5086", ROM_PC5086, "pc5086", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_PS2, 640, 640, 0, pc5086_init, &f82c710_upc_device};
-MODEL m_26 = {"[8086] Amstrad PPC512/640", ROM_PPC512, "ppc512", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_HW | MODEL_AMSTRAD, 512, 640, 128, ams_init, &ams1512_device};
-MODEL m_27 = {"[8086] Compaq Deskpro", ROM_DESKPRO, "deskpro", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 128, 640, 128, compaq_xt_init, NULL};
-MODEL m_28 = {"[8086] Olivetti M24", ROM_OLIM24, "olivetti_m24", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED | MODEL_OLIM24, 128, 640, 128, olim24_init, NULL};
-MODEL m_29 = {"[8086] Sinclair PC200", ROM_PC200, "pc200", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_HW | MODEL_AMSTRAD, 512, 640, 128, ams_init, &ams1512_device};
-MODEL m_30 = {"[8086] Tandy 1000 SL/2", ROM_TANDY1000SL2, "tandy1000sl2", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED, 512, 768, 128, tandy1ksl2_init, NULL};
-MODEL m_31 = {"[8088] Toshiba T1200", ROM_T1200, "t1200", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED, 1024, 2048, 1024, xt_t1200_init, NULL};
-MODEL m_32 = {"[8086] VTech Laser XT3", ROM_LXT3, "lxt3", {{"", cpus_8086}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE, 512, 1152, 128, xt_laserxt_init, NULL};
+/* 286 PC's */
+MODEL m_ami286 = {"[286] AMI 286 clone", ROM_AMI286, "ami286", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 512, 16384, 128, at_neat_init, NULL};
+MODEL m_award286 = {"[286] Award 286 clone", ROM_AWARD286, "award286", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 512, 16384, 128, at_scat_init, NULL};
+MODEL m_bull_micral_45 = {"[286] Bull Micral 45", ROM_BULL_MICRAL_45, "bull_micral_45", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 1024, 6144, 128, ibm_at_init, NULL};
+MODEL m_cmdpc30 = {"[286] Commodore PC 30 III", ROM_CMDPC30, "cmdpc30", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 640, 16384, 128, at_cbm_init, NULL};
+MODEL m_compaq_pii = {"[286] Compaq Portable II", ROM_COMPAQ_PII, "compaq_pii", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 256, 15872, 128, ibm_at_init, NULL};
+MODEL m_dells200 = {"[286] DELL System 200", ROM_DELL200, "dells200", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT, 640, 16384, 128, dells200_init, NULL};
+MODEL m_epson_pcax = {"[286] Epson PC AX", ROM_EPSON_PCAX, "epson_pcax", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT, 256, 15872, 128, at_init, NULL};
+MODEL m_epson_pcax2e = {"[286] Epson PC AX2e", ROM_EPSON_PCAX2E, "epson_pcax2e", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PS2, 256, 15872, 128, at_init, NULL};
+MODEL m_gdc212m = {"[286] Goldstar GDC-212M", ROM_GDC212M, "gdc212m", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 512, 4096, 512, at_scat_init, NULL};
+MODEL m_gw286ct = {"[286] GW-286CT GEAR", ROM_GW286CT, "gw286ct", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT, 512, 16384, 128, at_scat_init, NULL};
+MODEL m_super286tr = {"[286] Hyundai Super-286TR", ROM_HYUNDAI_SUPER286TR, "super286tr", {{"AMD", cpus_super286tr}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 1024, 4096, 128, at_scat_init, &f82c710_upc_device};
+MODEL m_ibmat = {"[286] IBM AT", ROM_IBMAT, "ibmat", {{"", cpus_ibmat}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT, 256, 15872, 128, ibm_at_init, NULL};
+MODEL m_ibmps1es = {"[286] IBM PS/1 model 2011", ROM_IBMPS1_2011, "ibmps1es", {{"", cpus_ps1_m2011}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED | MODEL_AT | MODEL_PS2, 512, 16384, 512, ps1_m2011_init, NULL};
+MODEL m_ibmps2_m30_286 = {"[286] IBM PS/2 Model 30-286", ROM_IBMPS2_M30_286, "ibmps2_m30_286", {{"", cpus_ps2_m30_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED | MODEL_AT | MODEL_PS2, 1, 16, 1, ps2_m30_286_init, NULL};
+MODEL m_ibmps2_m50 = {"[286] IBM PS/2 Model 50", ROM_IBMPS2_M50, "ibmps2_m50", {{"", cpus_ps2_m30_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PS2 | MODEL_MCA, 1, 16, 1, ps2_model_50_init, NULL};
+MODEL m_ibmxt286 = {"[286] IBM XT Model 286", ROM_IBMXT286, "ibmxt286", {{"", cpus_ibmxt286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT, 256, 15872, 128, ibm_at_init, NULL};
+MODEL m_spc4200p = {"[286] Samsung SPC-4200P", ROM_SPC4200P, "spc4200p", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 512, 2048, 128, at_scat_init, NULL};
+MODEL m_spc4216p = {"[286] Samsung SPC-4216P", ROM_SPC4216P, "spc4216p", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 1, 5, 1, at_scat_init, NULL};
+MODEL m_spc4620p = {"[286] Samsung SPC-4620P", ROM_SPC4620P, "spc4620p", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_HW | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 1, 5, 1, at_scat_init, NULL};
+MODEL m_t3100e = {"[286] Toshiba T3100e", ROM_T3100E, "t3100e", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED | MODEL_AT | MODEL_HAS_IDE, 1024, 5120, 256, at_t3100e_init, NULL};
+MODEL m_tg286m = {"[286] Trigem 286M", ROM_TG286M, "tg286m", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 512, 8192, 128, at_headland_init, NULL};
+MODEL m_tulip_tc7 = {"[286] Tulip AT Compact", ROM_TULIP_TC7, "tulip_tc7", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 640, 15872, 128, ibm_at_init, NULL};
 
-MODEL m_33 = {"[286] AMI 286 clone", ROM_AMI286, "ami286", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 512, 16384, 128, at_neat_init, NULL};
-MODEL m_34 = {"[286] Award 286 clone", ROM_AWARD286, "award286", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 512, 16384, 128, at_scat_init, NULL};
-MODEL m_35 = {"[286] Bull Micral 45", ROM_BULL_MICRAL_45, "bull_micral_45", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 1024, 6144, 128, ibm_at_init, NULL};
-MODEL m_36 = {"[286] Commodore PC 30 III", ROM_CMDPC30, "cmdpc30", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 640, 16384, 128, at_cbm_init, NULL};
-MODEL m_37 = {"[286] Compaq Portable II", ROM_COMPAQ_PII, "compaq_pii", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 256, 15872, 128, ibm_at_init, NULL};
-MODEL m_38 = {"[286] DELL System 200", ROM_DELL200, "dells200", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT, 640, 16384, 128, dells200_init, NULL};
-MODEL m_39 = {"[286] Epson PC AX", ROM_EPSON_PCAX, "epson_pcax", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT, 256, 15872, 128, at_init, NULL};
-MODEL m_40 = {"[286] Epson PC AX2e", ROM_EPSON_PCAX2E, "epson_pcax2e", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PS2, 256, 15872, 128, at_init, NULL};
-MODEL m_41 = {"[286] Goldstar GDC-212M", ROM_GDC212M, "gdc212m", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 512, 4096, 512, at_scat_init, NULL};
-MODEL m_42 = {"[286] GW-286CT GEAR", ROM_GW286CT, "gw286ct", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT, 512, 16384, 128, at_scat_init, NULL};
-MODEL m_43 = {"[286] Hyundai Super-286TR", ROM_HYUNDAI_SUPER286TR, "super286tr", {{"AMD", cpus_super286tr}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 1024, 4096, 128, at_scat_init, &f82c710_upc_device};
-MODEL m_44 = {"[286] IBM AT", ROM_IBMAT, "ibmat", {{"", cpus_ibmat}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT, 256, 15872, 128, ibm_at_init, NULL};
-MODEL m_45 = {"[286] IBM PS/1 model 2011", ROM_IBMPS1_2011, "ibmps1es", {{"", cpus_ps1_m2011}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED | MODEL_AT | MODEL_PS2, 512, 16384, 512, ps1_m2011_init, NULL};
-MODEL m_46 = {"[286] IBM PS/2 Model 30-286", ROM_IBMPS2_M30_286, "ibmps2_m30_286", {{"", cpus_ps2_m30_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED | MODEL_AT | MODEL_PS2, 1, 16, 1, ps2_m30_286_init, NULL};
-MODEL m_47 = {"[286] IBM PS/2 Model 50", ROM_IBMPS2_M50, "ibmps2_m50", {{"", cpus_ps2_m30_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PS2 | MODEL_MCA, 1, 16, 1, ps2_model_50_init, NULL};
-MODEL m_48 = {"[286] IBM XT Model 286", ROM_IBMXT286, "ibmxt286", {{"", cpus_ibmxt286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT, 256, 15872, 128, ibm_at_init, NULL};
-MODEL m_49 = {"[286] Samsung SPC-4200P", ROM_SPC4200P, "spc4200p", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 512, 2048, 128, at_scat_init, NULL};
-MODEL m_50 = {"[286] Samsung SPC-4216P", ROM_SPC4216P, "spc4216p", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 1, 5, 1, at_scat_init, NULL};
-MODEL m_51 = {"[286] Samsung SPC-4620P", ROM_SPC4620P, "spc4620p", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_HW | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 1, 5, 1, at_scat_init, NULL};
-MODEL m_52 = {"[286] Toshiba T3100e", ROM_T3100E, "t3100e", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_FIXED | MODEL_AT | MODEL_HAS_IDE, 1024, 5120, 256, at_t3100e_init, NULL};
-MODEL m_53 = {"[286] Trigem 286M", ROM_TG286M, "tg286m", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 512, 8192, 128, at_headland_init, NULL};
-MODEL m_54 = {"[286] Tulip AT Compact", ROM_TULIP_TC7, "tulip_tc7", {{"", cpus_286}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 640, 15872, 128, ibm_at_init, NULL};
+/* 386SX PC's */
+MODEL m_acer386 = {"[386SX] Acer 386SX25/N", ROM_ACER386, "acer386", {{"Intel", cpus_acer}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 1, 16, 1, at_acer386sx_init, NULL};
+MODEL m_ama932j = {"[386SX] AMA-932J", ROM_AMA932J, "ama932j", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_FIXED | MODEL_AT | MODEL_HAS_IDE, 512, 8192, 128, at_headland_init, NULL};
+MODEL m_ami386 = {"[386SX] AMI 386SX clone", ROM_AMI386SX, "ami386", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 512, 16384, 128, at_headland_init, NULL};
+MODEL m_megapc = {"[386SX] Amstrad MegaPC", ROM_MEGAPC, "megapc", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_DISABLE_HW | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 1, 16, 1, at_wd76c10_init, NULL};
+MODEL m_cbm_sl386sx25 = {"[386SX] Commodore SL386SX-25", ROM_CBM_SL386SX25, "cbm_sl386sx25", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_FIXED | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 1024, 16384, 512, at_scamp_init, NULL};
+MODEL m_dtk386 = {"[386SX] DTK 386SX clone", ROM_DTK386, "dtk386", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 512, 16384, 128, at_neat_init, NULL};
+MODEL m_epson_pcax3 = {"[386SX] Epson PC AX3", ROM_EPSON_PCAX3, "epson_pcax3", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_NONE | MODEL_AT, 256, 15872, 128, at_init, NULL};
+MODEL m_ibmps1_2121 = {"[386SX] IBM PS/1 model 2121", ROM_IBMPS1_2121, "ibmps1_2121", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_FIXED | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 1, 16, 1, ps1_m2121_init, NULL};
+MODEL m_ibmps2_m55sx = {"[386SX] IBM PS/2 Model 55SX", ROM_IBMPS2_M55SX, "ibmps2_m55sx", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PS2 | MODEL_MCA, 1, 8, 1, ps2_model_55sx_init, NULL};
+MODEL m_kmxc02 = {"[386SX] KMX-C-02", ROM_KMXC02, "kmxc02", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_NONE | MODEL_AT, 512, 16384, 512, at_scatsx_init, NULL};
+MODEL m_pb_l300sx = {"[386SX] Packard Bell Legend 300SX", ROM_PB_L300SX, "pb_l300sx", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 1, 16, 1, pb_l300sx_init, NULL};
+MODEL m_spc6033p = {"[386SX] Samsung SPC-6033P", ROM_SPC6033P, "spc6033p", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_DISABLE_HW | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 2, 12, 2, at_scamp_init, NULL};
 
-MODEL m_55 = {"[386SX] Acer 386SX25/N", ROM_ACER386, "acer386", {{"Intel", cpus_acer}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 1, 16, 1, at_acer386sx_init, NULL};
-MODEL m_56 = {"[386SX] AMA-932J", ROM_AMA932J, "ama932j", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_FIXED | MODEL_AT | MODEL_HAS_IDE, 512, 8192, 128, at_headland_init, NULL};
-MODEL m_57 = {"[386SX] AMI 386SX clone", ROM_AMI386SX, "ami386", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 512, 16384, 128, at_headland_init, NULL};
-MODEL m_58 = {"[386SX] Amstrad MegaPC", ROM_MEGAPC, "megapc", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_DISABLE_HW | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 1, 16, 1, at_wd76c10_init, NULL};
-MODEL m_59 = {"[386SX] Commodore SL386SX-25", ROM_CBM_SL386SX25, "cbm_sl386sx25", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_FIXED | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 1024, 16384, 512, at_scamp_init, NULL};
-MODEL m_60 = {"[386SX] DTK 386SX clone", ROM_DTK386, "dtk386", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 512, 16384, 128, at_neat_init, NULL};
-MODEL m_61 = {"[386SX] Epson PC AX3", ROM_EPSON_PCAX3, "epson_pcax3", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_NONE | MODEL_AT, 256, 15872, 128, at_init, NULL};
-MODEL m_62 = {"[386SX] IBM PS/1 model 2121", ROM_IBMPS1_2121, "ibmps1_2121", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_FIXED | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 1, 16, 1, ps1_m2121_init, NULL};
-MODEL m_63 = {"[386SX] IBM PS/2 Model 55SX", ROM_IBMPS2_M55SX, "ibmps2_m55sx", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PS2 | MODEL_MCA, 1, 8, 1, ps2_model_55sx_init, NULL};
-MODEL m_64 = {"[386SX] KMX-C-02", ROM_KMXC02, "kmxc02", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_NONE | MODEL_AT, 512, 16384, 512, at_scatsx_init, NULL};
-MODEL m_65 = {"[386SX] Packard Bell Legend 300SX", ROM_PB_L300SX, "pb_l300sx", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 1, 16, 1, pb_l300sx_init, NULL};
-MODEL m_66 = {"[386SX] Samsung SPC-6033P", ROM_SPC6033P, "spc6033p", {{"Intel", cpus_i386SX}, {"AMD", cpus_Am386SX}, {"Cyrix", cpus_486SLC}}, MODEL_GFX_DISABLE_HW | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 2, 12, 2, at_scamp_init, NULL};
+/* 386DX PC's */
+MODEL m_ami386dx = {"[386DX] AMI 386DX clone", ROM_AMI386DX_OPTI495, "ami386dx", {{"Intel", cpus_i386DX}, {"AMD", cpus_Am386DX}, {"Cyrix", cpus_486DLC}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 1, 256, 1, at_opti495_init, NULL};
+MODEL m_deskpro386 = {"[386DX] Compaq Deskpro 386", ROM_DESKPRO_386, "deskpro386", {{"Intel", cpus_i386DX}, {"AMD", cpus_Am386DX}, {"Cyrix", cpus_486DLC}}, MODEL_GFX_NONE | MODEL_AT, 1, 15, 1, deskpro386_init, NULL};
+MODEL m_ecs_386_32 = {"[386DX] ECS 386/32", ROM_ECS_386_32, "ecs_386_32", {{"Intel", cpus_i386DX}, {"AMD", cpus_Am386DX}, {"Cyrix", cpus_486DLC}}, MODEL_GFX_NONE | MODEL_AT, 1, 16, 1, at_cs8230_init, NULL};
+MODEL m_ibmps2_m70_type3 = {"[386DX] IBM PS/2 Model 70 (type 3)", ROM_IBMPS2_M70_TYPE3, "ibmps2_m70_type3", {{"Intel", cpus_i386DX}, {"AMD", cpus_Am386DX}, {"Cyrix", cpus_486DLC}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PS2 | MODEL_MCA, 2, 16, 2, ps2_model_70_init, NULL};
+MODEL m_ibmps2_m80 = {"[386DX] IBM PS/2 Model 80", ROM_IBMPS2_M80, "ibmps2_m80", {{"Intel", cpus_i386DX}, {"AMD", cpus_Am386DX}, {"Cyrix", cpus_486DLC}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PS2 | MODEL_MCA, 1, 12, 1, ps2_model_80_init, NULL};
+MODEL m_mr386dx = {"[386DX] MR 386DX clone", ROM_MR386DX_OPTI495, "mr386dx", {{"Intel", cpus_i386DX}, {"AMD", cpus_Am386DX}, {"Cyrix", cpus_486DLC}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 1, 256, 1, at_opti495_init, NULL};
+MODEL m_spc6000a = {"[386DX] Samsung SPC-6000A", ROM_SPC6000A, "spc6000a", {{"Intel", cpus_i386DX}, {"AMD", cpus_Am386DX}, {"Cyrix", cpus_486DLC}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 1, 32, 1, at_cs8230_init, NULL};
 
-MODEL m_67 = {"[386DX] AMI 386DX clone", ROM_AMI386DX_OPTI495, "ami386dx", {{"Intel", cpus_i386DX}, {"AMD", cpus_Am386DX}, {"Cyrix", cpus_486DLC}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 1, 256, 1, at_opti495_init, NULL};
-MODEL m_68 = {"[386DX] Compaq Deskpro 386", ROM_DESKPRO_386, "deskpro386", {{"Intel", cpus_i386DX}, {"AMD", cpus_Am386DX}, {"Cyrix", cpus_486DLC}}, MODEL_GFX_NONE | MODEL_AT, 1, 15, 1, deskpro386_init, NULL};
-MODEL m_69 = {"[386DX] ECS 386/32", ROM_ECS_386_32, "ecs_386_32", {{"Intel", cpus_i386DX}, {"AMD", cpus_Am386DX}, {"Cyrix", cpus_486DLC}}, MODEL_GFX_NONE | MODEL_AT, 1, 16, 1, at_cs8230_init, NULL};
-MODEL m_70 = {"[386DX] IBM PS/2 Model 70 (type 3)", ROM_IBMPS2_M70_TYPE3, "ibmps2_m70_type3", {{"Intel", cpus_i386DX}, {"AMD", cpus_Am386DX}, {"Cyrix", cpus_486DLC}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PS2 | MODEL_MCA, 2, 16, 2, ps2_model_70_init, NULL};
-MODEL m_71 = {"[386DX] IBM PS/2 Model 80", ROM_IBMPS2_M80, "ibmps2_m80", {{"Intel", cpus_i386DX}, {"AMD", cpus_Am386DX}, {"Cyrix", cpus_486DLC}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PS2 | MODEL_MCA, 1, 12, 1, ps2_model_80_init, NULL};
-MODEL m_72 = {"[386DX] MR 386DX clone", ROM_MR386DX_OPTI495, "mr386dx", {{"Intel", cpus_i386DX}, {"AMD", cpus_Am386DX}, {"Cyrix", cpus_486DLC}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 1, 256, 1, at_opti495_init, NULL};
-MODEL m_73 = {"[386DX] Samsung SPC-6000A", ROM_SPC6000A, "spc6000a", {{"Intel", cpus_i386DX}, {"AMD", cpus_Am386DX}, {"Cyrix", cpus_486DLC}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 1, 32, 1, at_cs8230_init, NULL};
+/* 486 PC's */
+MODEL m_ami486 = {"[486] AMI 486 clone", ROM_AMI486, "ami486", {{"Intel", cpus_i486}, {"AMD", cpus_Am486}, {"Cyrix", cpus_Cx486}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 1, 256, 1, at_ali1429_init, NULL};
+MODEL m_win486 = {"[486] AMI WinBIOS 486", ROM_WIN486, "win486", {{"Intel", cpus_i486}, {"AMD", cpus_Am486}, {"Cyrix", cpus_Cx486}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 1, 256, 1, at_ali1429_init, NULL};
+MODEL m_sis496 = {"[486] Award SiS 496/497", ROM_SIS496, "sis496", {{"Intel", cpus_i486}, {"AMD", cpus_Am486}, {"Cyrix", cpus_Cx486}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_HAS_IDE, 1, 256, 1, at_sis496_init, NULL};
+MODEL m_elx_pc425x = {"[486] Elonex PC-425X", ROM_ELX_PC425X, "elx_pc425x", {{"Intel", cpus_i486}, {"AMD", cpus_Am486}, {"Cyrix", cpus_Cx486}}, MODEL_GFX_FIXED | MODEL_AT | MODEL_HAS_IDE, 1, 256, 1, at_sl82c460_init, NULL};
+MODEL m_ibmps1_2133 = {"[486] IBM PS/1 Model 2133 (EMEA 451)", ROM_IBMPS1_2133_451, "ibmps1_2133", {{"Intel", cpus_i486}, {"AMD", cpus_Am486}, {"Cyrix", cpus_Cx486}}, MODEL_GFX_FIXED | MODEL_AT | MODEL_PS2, 2, 64, 2, ps1_m2133_init, NULL};
+MODEL m_ibmps2_m70_type4 = {"[486] IBM PS/2 Model 70 (type 4)", ROM_IBMPS2_M70_TYPE4, "ibmps2_m70_type4", {{"Intel", cpus_i486}, {"AMD", cpus_Am486}, {"Cyrix", cpus_Cx486}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PS2 | MODEL_MCA, 2, 16, 2, ps2_model_70_init, NULL};
+MODEL m_pb410a = {"[486] Packard Bell PB410A", ROM_PB410A, "pb410a", {{"Intel", cpus_i486}, {"AMD", cpus_Am486}, {"Cyrix", cpus_Cx486}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 1, 64, 1, at_pb410a_init, NULL};
 
-MODEL m_74 = {"[486] AMI 486 clone", ROM_AMI486, "ami486", {{"Intel", cpus_i486}, {"AMD", cpus_Am486}, {"Cyrix", cpus_Cx486}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 1, 256, 1, at_ali1429_init, NULL};
-MODEL m_75 = {"[486] AMI WinBIOS 486", ROM_WIN486, "win486", {{"Intel", cpus_i486}, {"AMD", cpus_Am486}, {"Cyrix", cpus_Cx486}}, MODEL_GFX_NONE | MODEL_AT | MODEL_HAS_IDE, 1, 256, 1, at_ali1429_init, NULL};
-MODEL m_76 = {"[486] Award SiS 496/497", ROM_SIS496, "sis496", {{"Intel", cpus_i486}, {"AMD", cpus_Am486}, {"Cyrix", cpus_Cx486}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_HAS_IDE, 1, 256, 1, at_sis496_init, NULL};
-MODEL m_77 = {"[486] Elonex PC-425X", ROM_ELX_PC425X, "elx_pc425x", {{"Intel", cpus_i486}, {"AMD", cpus_Am486}, {"Cyrix", cpus_Cx486}}, MODEL_GFX_FIXED | MODEL_AT | MODEL_HAS_IDE, 1, 256, 1, at_sl82c460_init, NULL};
-MODEL m_78 = {"[486] IBM PS/1 Model 2133 (EMEA 451)", ROM_IBMPS1_2133_451, "ibmps1_2133", {{"Intel", cpus_i486}, {"AMD", cpus_Am486}, {"Cyrix", cpus_Cx486}}, MODEL_GFX_FIXED | MODEL_AT | MODEL_PS2, 2, 64, 2, ps1_m2133_init, NULL};
-MODEL m_79 = {"[486] IBM PS/2 Model 70 (type 4)", ROM_IBMPS2_M70_TYPE4, "ibmps2_m70_type4", {{"Intel", cpus_i486}, {"AMD", cpus_Am486}, {"Cyrix", cpus_Cx486}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PS2 | MODEL_MCA, 2, 16, 2, ps2_model_70_init, NULL};
-MODEL m_80 = {"[486] Packard Bell PB410A", ROM_PB410A, "pb410a", {{"Intel", cpus_i486}, {"AMD", cpus_Am486}, {"Cyrix", cpus_Cx486}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PS2 | MODEL_HAS_IDE, 1, 64, 1, at_pb410a_init, NULL};
+/* Socket 4 PC's */
+MODEL m_revenge = {"[Socket 4] Intel Premiere/PCI", ROM_REVENGE, "revenge", {{"Intel", cpus_Pentium5V}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 1, 128, 1, at_batman_init, NULL};
+MODEL m_pb520r = {"[Socket 4] Packard Bell PB520R", ROM_PB520R, "pb520r", {{"Intel", cpus_Pentium5V}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 1, 128, 1, at_pb520r_init, NULL};
 
-MODEL m_81 = {"[Socket 4] Intel Premiere/PCI", ROM_REVENGE, "revenge", {{"Intel", cpus_Pentium5V}, {"", NULL}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 1, 128, 1, at_batman_init, NULL};
-MODEL m_82 = {"[Socket 4] Packard Bell PB520R", ROM_PB520R, "pb520r", {{"Intel", cpus_Pentium5V}, {"", NULL}, {"", NULL}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 1, 128, 1, at_pb520r_init, NULL};
+/* Socket 5 PC's */
+MODEL m_endeavor = {"[Socket 5] Intel Advanced/EV", ROM_ENDEAVOR, "endeavor", {{"Intel", cpus_PentiumS5}, {"IDT", cpus_WinChip}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 1, 128, 1, at_endeavor_init, NULL};
+MODEL m_zappa = {"[Socket 5] Intel Advanced/ZP", ROM_ZAPPA, "zappa", {{"Intel", cpus_PentiumS5}, {"IDT", cpus_WinChip}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 1, 128, 1, at_zappa_init, NULL};
+MODEL m_infowaym = {"[Socket 5] Itautec Infoway Multimidia", ROM_ITAUTEC_INFOWAYM, "infowaym", {{"Intel", cpus_PentiumS5}, {"IDT", cpus_WinChip}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 8, 128, 1, at_zappa_init, NULL};
+MODEL m_pb570 = {"[Socket 5] Packard Bell PB570", ROM_PB570, "pb570", {{"Intel", cpus_PentiumS5}, {"IDT", cpus_WinChip}, {"", NULL}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 1, 128, 1, at_pb570_init, NULL};
 
-MODEL m_83 = {"[Socket 5] Intel Advanced/EV", ROM_ENDEAVOR, "endeavor", {{"Intel", cpus_PentiumS5}, {"IDT", cpus_WinChip}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 1, 128, 1, at_endeavor_init, NULL};
-MODEL m_84 = {"[Socket 5] Intel Advanced/ZP", ROM_ZAPPA, "zappa", {{"Intel", cpus_PentiumS5}, {"IDT", cpus_WinChip}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 1, 128, 1, at_zappa_init, NULL};
-MODEL m_85 = {"[Socket 5] Itautec Infoway Multimidia", ROM_ITAUTEC_INFOWAYM, "infowaym", {{"Intel", cpus_PentiumS5}, {"IDT", cpus_WinChip}, {"", NULL}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 8, 128, 1, at_zappa_init, NULL};
-MODEL m_86 = {"[Socket 5] Packard Bell PB570", ROM_PB570, "pb570", {{"Intel", cpus_PentiumS5}, {"IDT", cpus_WinChip}, {"", NULL}}, MODEL_GFX_DISABLE_SW | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 1, 128, 1, at_pb570_init, NULL};
+/* Socket 7 PC's */
+MODEL m_p55tvp4 = {"[Socket 7] ASUS P/I-P55TVP4", ROM_P55TVP4, "p55tvp4", {{"Intel", cpus_Pentium}, {"AMD", cpus_K6_S7}, {"IDT", cpus_WinChip}, {"Cyrix", cpus_6x86}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 8, 128, 1, at_p55tvp4_init, NULL};
+MODEL m_p55t2p4 = {"[Socket 7] ASUS P/I-P55T2P4", ROM_P55T2P4, "p55t2p4", {{"Intel", cpus_Pentium}, {"AMD", cpus_K6_S7}, {"IDT", cpus_WinChip}, {"Cyrix", cpus_6x86}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 8, 256, 1, at_p55t2p4_init, NULL};
+MODEL m_p55va = {"[Socket 7] Epox P55-VA", ROM_P55VA, "p55va", {{"Intel", cpus_Pentium}, {"AMD", cpus_K6_S7}, {"IDT", cpus_WinChip}, {"Cyrix", cpus_6x86}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 8, 128, 1, at_p55va_init, NULL};
+MODEL m_430vx = {"[Socket 7] Shuttle HOT-557", ROM_430VX, "430vx", {{"Intel", cpus_Pentium}, {"AMD", cpus_K6_S7}, {"IDT", cpus_WinChip}, {"Cyrix", cpus_6x86}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 8, 128, 1, at_i430vx_init, NULL};
 
-MODEL m_87 = {"[Socket 7] ASUS P/I-P55TVP4", ROM_P55TVP4, "p55tvp4", {{"Intel", cpus_Pentium}, {"AMD", cpus_K6_S7}, {"IDT", cpus_WinChip}, {"Cyrix", cpus_6x86}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 8, 128, 1, at_p55tvp4_init, NULL};
-MODEL m_88 = {"[Socket 7] ASUS P/I-P55T2P4", ROM_P55T2P4, "p55t2p4", {{"Intel", cpus_Pentium}, {"AMD", cpus_K6_S7}, {"IDT", cpus_WinChip}, {"Cyrix", cpus_6x86}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 8, 256, 1, at_p55t2p4_init, NULL};
-MODEL m_89 = {"[Socket 7] Epox P55-VA", ROM_P55VA, "p55va", {{"Intel", cpus_Pentium}, {"AMD", cpus_K6_S7}, {"IDT", cpus_WinChip}, {"Cyrix", cpus_6x86}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 8, 128, 1, at_p55va_init, NULL};
-MODEL m_90 = {"[Socket 7] Shuttle HOT-557", ROM_430VX, "430vx", {{"Intel", cpus_Pentium}, {"AMD", cpus_K6_S7}, {"IDT", cpus_WinChip}, {"Cyrix", cpus_6x86}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 8, 128, 1, at_i430vx_init, NULL};
+/* Super 7 PC's */
+MODEL m_fic_va503p = {"[Super 7] FIC VA-503+", ROM_FIC_VA503P, "fic_va503p", {{"Intel", cpus_Pentium}, {"AMD", cpus_K6_SS7}, {"IDT", cpus_WinChip_SS7}, {"Cyrix", cpus_6x86_SS7}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 1, 512, 1, at_mvp3_init, NULL};
 
-MODEL m_91 = {"[Super 7] FIC VA-503+", ROM_FIC_VA503P, "fic_va503p", {{"Intel", cpus_Pentium}, {"AMD", cpus_K6_SS7}, {"IDT", cpus_WinChip_SS7}, {"Cyrix", cpus_6x86_SS7}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 1, 512, 1, at_mvp3_init, NULL};
+/* Socket 8 PC's */
+MODEL m_vs440fx = {"[Socket 8] Intel VS440FX", ROM_VS440FX, "vs440fx", {{"Intel", cpus_PentiumPro}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 8, 256, 8, at_vs440fx_init, NULL};
 
-MODEL m_92 = {"[Socket 8] Intel VS440FX", ROM_VS440FX, "vs440fx", {{"Intel", cpus_PentiumPro}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 8, 256, 8, at_vs440fx_init, NULL};
-
-MODEL m_93 = {"[Slot 1] Gigabyte GA-686BX", ROM_GA686BX, "ga686bx", {{"Intel", cpus_Slot1_100MHz}, {"VIA", cpus_VIA_100MHz}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 8, 512, 8, at_ga686bx_init, NULL};
+/* Slot 1 PC's */
+MODEL m_ga686bx = {"[Slot 1] Gigabyte GA-686BX", ROM_GA686BX, "ga686bx", {{"Intel", cpus_Slot1_100MHz}, {"VIA", cpus_VIA_100MHz}}, MODEL_GFX_NONE | MODEL_AT | MODEL_PCI | MODEL_PS2 | MODEL_HAS_IDE, 8, 512, 8, at_ga686bx_init, NULL};
 
 void pcem_add_model(MODEL *model)
 {
+        //TODO: Add sanity check to not go past MAX amount
         models[model_count()] = model;
 }
 
@@ -895,100 +907,123 @@ void model_init_builtin()
 {
         memset(models, 0, sizeof(models));
 
-        pcem_add_model(&m_1);
-        pcem_add_model(&m_2);
-        pcem_add_model(&m_cbmPC10);
-        pcem_add_model(&m_3);
-        pcem_add_model(&m_4);
-        pcem_add_model(&m_5);
-        pcem_add_model(&m_huyndai16t);
-        pcem_add_model(&m_huyndai16te);
-        pcem_add_model(&m_6);
-        pcem_add_model(&m_7);
-        pcem_add_model(&m_8);
-        pcem_add_model(&m_9);
-        pcem_add_model(&m_10);
-        pcem_add_model(&m_11);
-        pcem_add_model(&m_12);
-        pcem_add_model(&m_13);
-        pcem_add_model(&m_14);
-        pcem_add_model(&m_15);
-        pcem_add_model(&m_16);
-        pcem_add_model(&m_17);
-        pcem_add_model(&m_18);
-        pcem_add_model(&m_19);
-        pcem_add_model(&m_20);
-        pcem_add_model(&m_21);
-        pcem_add_model(&m_22);
-        pcem_add_model(&m_23);
-        pcem_add_model(&m_24);
-        pcem_add_model(&m_25);
-        pcem_add_model(&m_26);
-        pcem_add_model(&m_27);
-        pcem_add_model(&m_28);
-        pcem_add_model(&m_29);
-        pcem_add_model(&m_30);
-        pcem_add_model(&m_31);
-        pcem_add_model(&m_32);
-        pcem_add_model(&m_33);
-        pcem_add_model(&m_34);
-        pcem_add_model(&m_35);
-        pcem_add_model(&m_36);
-        pcem_add_model(&m_37);
-        pcem_add_model(&m_38);
-        pcem_add_model(&m_39);
-        pcem_add_model(&m_40);
-        pcem_add_model(&m_41);
-        pcem_add_model(&m_42);
-        pcem_add_model(&m_43);
-        pcem_add_model(&m_44);
-        pcem_add_model(&m_45);
-        pcem_add_model(&m_46);
-        pcem_add_model(&m_47);
-        pcem_add_model(&m_48);
-        pcem_add_model(&m_49);
-        pcem_add_model(&m_50);
-        pcem_add_model(&m_51);
-        pcem_add_model(&m_52);
-        pcem_add_model(&m_53);
-        pcem_add_model(&m_54);
-        pcem_add_model(&m_55);
-        pcem_add_model(&m_56);
-        pcem_add_model(&m_57);
-        pcem_add_model(&m_58);
-        pcem_add_model(&m_59);
-        pcem_add_model(&m_60);
-        pcem_add_model(&m_61);
-        pcem_add_model(&m_62);
-        pcem_add_model(&m_63);
-        pcem_add_model(&m_64);
-        pcem_add_model(&m_65);
-        pcem_add_model(&m_66);
-        pcem_add_model(&m_67);
-        pcem_add_model(&m_68);
-        pcem_add_model(&m_69);
-        pcem_add_model(&m_70);
-        pcem_add_model(&m_71);
-        pcem_add_model(&m_72);
-        pcem_add_model(&m_73);
-        pcem_add_model(&m_74);
-        pcem_add_model(&m_75);
-        pcem_add_model(&m_76);
-        pcem_add_model(&m_77);
-        pcem_add_model(&m_78);
-        pcem_add_model(&m_79);
-        pcem_add_model(&m_80);
-        pcem_add_model(&m_81);
-        pcem_add_model(&m_82);
-        pcem_add_model(&m_83);
-        pcem_add_model(&m_84);
-        pcem_add_model(&m_85);
-        pcem_add_model(&m_86);
-        pcem_add_model(&m_87);
-        pcem_add_model(&m_88);
-        pcem_add_model(&m_89);
-        pcem_add_model(&m_90);
-        pcem_add_model(&m_91);
-        pcem_add_model(&m_92);
-        pcem_add_model(&m_93);
+        /* 8088 PC's */
+        pcem_add_model(&m_amixt);
+        pcem_add_model(&m_ataripc3);
+        pcem_add_model(&m_cbm_pc10);
+        pcem_add_model(&m_compaq_pip);
+        pcem_add_model(&m_dtk);
+        pcem_add_model(&m_genxt);
+        pcem_add_model(&m_hyundaixt);
+        pcem_add_model(&m_hyundaixte);
+        pcem_add_model(&m_ibmpc);
+        pcem_add_model(&m_ibmpcjr);
+        pcem_add_model(&m_ibmxt);
+        pcem_add_model(&m_jukopc);
+        pcem_add_model(&m_ledge_modelm);
+        pcem_add_model(&m_ncr_pc4i);
+        pcem_add_model(&m_pxxt);
+        pcem_add_model(&m_europc);
+        pcem_add_model(&m_tandy);
+        pcem_add_model(&m_tandy1000hx);
+        pcem_add_model(&m_to16_pc);
+        pcem_add_model(&m_t1000);
+        pcem_add_model(&m_ltxt);
+        pcem_add_model(&m_xi8088);
+        pcem_add_model(&m_zdsupers);
+
+        /* 8086 PC's */
+        pcem_add_model(&m_pc1512);
+        pcem_add_model(&m_pc1640);
+        pcem_add_model(&m_pc2086);
+        pcem_add_model(&m_pc3086);
+        pcem_add_model(&m_pc5086);
+        pcem_add_model(&m_ppc512);
+        pcem_add_model(&m_deskpro);
+        pcem_add_model(&m_olivetti_m24);
+        pcem_add_model(&m_pc200);
+        pcem_add_model(&m_tandy1000sl2);
+        pcem_add_model(&m_t1200);
+        pcem_add_model(&m_lxt3);
+
+        /* 286 PC's */
+        pcem_add_model(&m_ami286);
+        pcem_add_model(&m_award286);
+        pcem_add_model(&m_bull_micral_45);
+        pcem_add_model(&m_cmdpc30);
+        pcem_add_model(&m_compaq_pii);
+        pcem_add_model(&m_dells200);
+        pcem_add_model(&m_epson_pcax);
+        pcem_add_model(&m_epson_pcax2e);
+        pcem_add_model(&m_gdc212m);
+        pcem_add_model(&m_gw286ct);
+        pcem_add_model(&m_super286tr);
+        pcem_add_model(&m_ibmat);
+        pcem_add_model(&m_ibmps1es);
+        pcem_add_model(&m_ibmps2_m30_286);
+        pcem_add_model(&m_ibmps2_m50);
+        pcem_add_model(&m_ibmxt286);
+        pcem_add_model(&m_spc4200p);
+        pcem_add_model(&m_spc4216p);
+        pcem_add_model(&m_spc4620p);
+        pcem_add_model(&m_t3100e);
+        pcem_add_model(&m_tg286m);
+        pcem_add_model(&m_tulip_tc7);
+
+        /* 386SX PC's */
+        pcem_add_model(&m_acer386);
+        pcem_add_model(&m_ama932j);
+        pcem_add_model(&m_ami386);
+        pcem_add_model(&m_megapc);
+        pcem_add_model(&m_cbm_sl386sx25);
+        pcem_add_model(&m_dtk386);
+        pcem_add_model(&m_epson_pcax3);
+        pcem_add_model(&m_ibmps1_2121);
+        pcem_add_model(&m_ibmps2_m55sx);
+        pcem_add_model(&m_kmxc02);
+        pcem_add_model(&m_pb_l300sx);
+        pcem_add_model(&m_spc6033p);
+
+        /* 386DX PC's */
+        pcem_add_model(&m_ami386dx);
+        pcem_add_model(&m_deskpro386);
+        pcem_add_model(&m_ecs_386_32);
+        pcem_add_model(&m_ibmps2_m70_type3);
+        pcem_add_model(&m_ibmps2_m80);
+        pcem_add_model(&m_mr386dx);
+        pcem_add_model(&m_spc6000a);
+
+        /* 486 PC's */
+        pcem_add_model(&m_ami486);
+        pcem_add_model(&m_win486);
+        pcem_add_model(&m_sis496);
+        pcem_add_model(&m_elx_pc425x);
+        pcem_add_model(&m_ibmps1_2133);
+        pcem_add_model(&m_ibmps2_m70_type4);
+        pcem_add_model(&m_pb410a);
+
+        /* Socket 4 PC's */
+        pcem_add_model(&m_revenge);
+        pcem_add_model(&m_pb520r);
+
+        /* Socket 5 PC's */
+        pcem_add_model(&m_endeavor);
+        pcem_add_model(&m_zappa);
+        pcem_add_model(&m_infowaym);
+        pcem_add_model(&m_pb570);
+
+        /* Socket 7 PC's */
+        pcem_add_model(&m_p55tvp4);
+        pcem_add_model(&m_p55t2p4);
+        pcem_add_model(&m_p55va);
+        pcem_add_model(&m_430vx);
+
+        /* Super 7 PC's */
+        pcem_add_model(&m_fic_va503p);
+
+        /* Socket 8 PC's */
+        pcem_add_model(&m_vs440fx);
+
+        /* Slot 1 PC's */
+        pcem_add_model(&m_ga686bx);
 }