+++ src/plugin.c
@@ -1,0 +1,6 @@
+#include "plugin.h"
+
+void init_plugin_engine()
+{
+
+}